Separate GUI functionality into mddgui.py
Moved all existing GUI-related code from mdd.py to a new file, mddgui.py. This includes the class definitions and related methods for managing the graphical interface. Made necessary imports and functions to ensure the GUI works independently.
